+++ v2/tools/asoctl/internal/importing/arm_resource_importer.go
@@ -13,29 +13,31 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/cloud"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
-	azruntime "github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
+	"github.com/Azure/azure-sdk-for-go/sdk/azidentity"
+	"github.com/Azure/azure-service-operator/v2/internal/genericarmclient"
 	"github.com/Azure/azure-service-operator/v2/internal/reflecthelpers"
+	"github.com/Azure/azure-service-operator/v2/internal/version"
 	"github.com/Azure/azure-service-operator/v2/pkg/genruntime"
 	"github.com/Azure/azure-service-operator/v2/tools/generator/pkg/naming"
 	"github.com/pkg/errors"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/klog/v2"
+
+	armruntime "github.com/Azure/azure-sdk-for-go/sdk/azcore/arm/runtime"
+	azruntime "github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
 )
 
-type ARMResourceImporter interface {
-	Import(ctx context.Context, armID string) (*resourceImportResult, error)
+// ARMResourceImporter is an importer for ARM based resources.
+// Create one using the factory method on ResourceImporter.
+type ARMResourceImporter struct {
+	ResourceImporter
+	client               *azruntime.Pipeline
+	serviceConfiguration cloud.ServiceConfiguration
 }
 
-type armResourceImporter struct {
-	resourceImporterFactory
-	armClient *azruntime.Pipeline
-	armConfig cloud.ServiceConfiguration
-}
-
-var _ ARMResourceImporter = &armResourceImporter{}
-
-func (ri *armResourceImporter) Import(ctx context.Context, armID string) (*resourceImportResult, error) {
+// Import imports the specified ARM resource, returning the imported resource.
+func (ri *ARMResourceImporter) Import(ctx context.Context, armID string) (*ResourceImportResult, error) {
 	// Parse armID into a more useful form
 	id, err := arm.ParseResourceID(armID)
 	if err != nil {
@@ -54,6 +56,23 @@ func (ri *armResourceImporter) Import(ctx context.Context, armID string) (*resou
 			"unable to create blank resource, expected %s to identify an ARM object", armID)
 	}
 
+	status, err := ri.getStatus(ctx, armID, armMeta)
+	if err != nil {
+		return nil, err
+	}
+
+	err = armMeta.SetStatus(status)
+	if err != nil {
+		return nil, errors.Wrapf(err, "setting status on Kubernetes resource for resource %s", armID)
+	}
+
+	return &ResourceImportResult{
+		resources: []genruntime.MetaObject{armMeta},
+	}, nil
+}
+
+// getStatus gets the status of the resource with the specified ID from ARM
+func (ri *ARMResourceImporter) getStatus(ctx context.Context, armID string, armMeta genruntime.ARMMetaObject) (genruntime.ConvertibleStatus, error) {
 	// Create a request to get the current state of the resource
 	req, err := ri.createRequest(ctx, armID, armMeta.GetAPIVersion())
 	if err != nil {
@@ -61,13 +80,12 @@ func (ri *armResourceImporter) Import(ctx context.Context, armID string) (*resou
 	}
 
 	// Execute the request
-	resp, err := ri.armClient.Do(req)
+	resp, err := ri.client.Do(req)
 	if err != nil {
 		return nil, errors.Wrapf(err, "unable to execute request to import ARM resource %s", armID)
 	}
 
 	if !azruntime.HasStatusCode(resp, http.StatusOK) {
-		klog.Warningf("Request failed with status code %d", resp.StatusCode)
 		return nil, azruntime.NewResponseError(resp)
 	}
 
@@ -77,16 +95,6 @@ func (ri *armResourceImporter) Import(ctx context.Context, armID string) (*resou
 	if err != nil {
 		return nil, errors.Wrapf(err, "unable to create empty ARM status for importing ARM resource %s", armID)
 	}
-
-	// Create an owner reference
-	var knownOwner genruntime.ArbitraryOwnerReference
-	//if owner := kr.Owner(); owner != nil {
-	//	knownOwner = genruntime.ArbitraryOwnerReference{
-	//		Name:  owner.Name,
-	//		Group: owner.Group,
-	//		Kind:  owner.Kind,
-	//	}
-	//}
 
 	// Populate our Status from the response
 	if err := azruntime.UnmarshalAsJSON(resp, armStatus); err != nil {
@@ -101,27 +109,20 @@ func (ri *armResourceImporter) Import(ctx context.Context, armID string) (*resou
 
 	//TODO: what if it's not ok
 	if s, ok := status.(genruntime.FromARMConverter); ok {
+		var knownOwner genruntime.ArbitraryOwnerReference
 		err = s.PopulateFromARM(knownOwner, reflecthelpers.ValueOfPtr(armStatus)) // TODO: PopulateFromArm expects a value... ick
 		if err != nil {
 			return nil, errors.Wrapf(err, "converting ARM status to Kubernetes status for resource %s", armID)
 		}
 	}
-
-	err = armMeta.SetStatus(status)
-	if err != nil {
-		return nil, errors.Wrapf(err, "setting status on Kubernetes resource for resource %s", armID)
-	}
-
-	return &resourceImportResult{
-		Object: armMeta,
-	}, nil
+	return status, nil
 }
 
 // createRequest constructs the request to GET the ARM resource
-func (ri *armResourceImporter) createRequest(ctx context.Context, armID string, apiVersion string) (*policy.Request, error) {
+func (ri *ARMResourceImporter) createRequest(ctx context.Context, armID string, apiVersion string) (*policy.Request, error) {
 	//urlPath = strings.ReplaceAll(urlPath, "{resourceId}", ari.armID)
 	//req, err := runtime.NewRequest(ctx, http.MethodGet, runtime.JoinPaths(rmConfig.Endpoint, urlPath))
-	req, err := azruntime.NewRequest(ctx, http.MethodGet, azruntime.JoinPaths(ri.armConfig.Endpoint, armID))
+	req, err := azruntime.NewRequest(ctx, http.MethodGet, azruntime.JoinPaths(ri.serviceConfiguration.Endpoint, armID))
 	if err != nil {
 		return nil, err
 	}
@@ -136,15 +137,15 @@ func (ri *armResourceImporter) createRequest(ctx context.Context, armID string, 
 	return req, nil
 }
 
-func (ri *armResourceImporter) Client() *azruntime.Pipeline {
-	return ri.armClient
+func (ri *ARMResourceImporter) Client() *azruntime.Pipeline {
+	return ri.client
 }
 
-func (ri *armResourceImporter) Config() cloud.ServiceConfiguration {
-	return ri.armConfig
+func (ri *ARMResourceImporter) Config() cloud.ServiceConfiguration {
+	return ri.serviceConfiguration
 }
 
-func (ri *armResourceImporter) createBlankObjectFromID(armID *arm.ResourceID) (runtime.Object, error) {
+func (ri *ARMResourceImporter) createBlankObjectFromID(armID *arm.ResourceID) (runtime.Object, error) {
 	gvk, err := ri.groupVersionKindFromID(armID)
 	if err != nil {
 		return nil, errors.Wrap(err, "unable to get GVK for blank resource")
@@ -168,7 +169,7 @@ func (ri *armResourceImporter) createBlankObjectFromID(armID *arm.ResourceID) (r
 }
 
 // groupVersionKindFromID returns the GroupVersionKind for the resource we're importing
-func (ri *armResourceImporter) groupVersionKindFromID(id *arm.ResourceID) (schema.GroupVersionKind, error) {
+func (ri *ARMResourceImporter) groupVersionKindFromID(id *arm.ResourceID) (schema.GroupVersionKind, error) {
 	gk, err := ri.groupKindFromID(id)
 	if err != nil {
 		return schema.GroupVersionKind{},
@@ -179,7 +180,7 @@ func (ri *armResourceImporter) groupVersionKindFromID(id *arm.ResourceID) (schem
 }
 
 // groupKindFromID parses a GroupKind from the resource URL, allowing us to look up the actual resource
-func (ri *armResourceImporter) groupKindFromID(id *arm.ResourceID) (schema.GroupKind, error) {
+func (ri *ARMResourceImporter) groupKindFromID(id *arm.ResourceID) (schema.GroupKind, error) {
 	return schema.GroupKind{
 		Group: ri.groupFromID(id),
 		Kind:  ri.kindFromID(id),
@@ -187,7 +188,7 @@ func (ri *armResourceImporter) groupKindFromID(id *arm.ResourceID) (schema.Group
 }
 
 // groupFromID extracts an ASO group name from the ARM ID
-func (*armResourceImporter) groupFromID(id *arm.ResourceID) string {
+func (*ARMResourceImporter) groupFromID(id *arm.ResourceID) string {
 	parts := strings.Split(id.ResourceType.Namespace, ".")
 	last := len(parts) - 1
 	group := strings.ToLower(parts[last]) + ".azure.com"
@@ -196,7 +197,7 @@ func (*armResourceImporter) groupFromID(id *arm.ResourceID) string {
 }
 
 // kindFromID extracts an ASO kind from the ARM ID
-func (*armResourceImporter) kindFromID(id *arm.ResourceID) string {
+func (*ARMResourceImporter) kindFromID(id *arm.ResourceID) string {
 	if len(id.ResourceType.Types) != 1 {
 		panic("Don't currently know how to handle nested resources")
 	}
@@ -206,7 +207,28 @@ func (*armResourceImporter) kindFromID(id *arm.ResourceID) string {
 	return kind
 }
 
-func (ri *armResourceImporter) nameFromID(id *arm.ResourceID) (string, error) {
+func (ri *ARMResourceImporter) nameFromID(id *arm.ResourceID) (string, error) {
 	klog.V(3).Infof("Name: %s", id.Name)
 	return id.Name, nil
+}
+
+func CreateARMClient(cloudConfig cloud.Configuration) (*azruntime.Pipeline, error) {
+	creds, err := azidentity.NewDefaultAzureCredential(nil)
+	if err != nil {
+		return nil, errors.Wrap(err, "unable to get default azure credential")
+	}
+
+	var userAgent = "asoctl/" + version.BuildVersion
+
+	opts := &arm.ClientOptions{
+		ClientOptions: policy.ClientOptions{
+			Cloud: cloudConfig,
+			PerCallPolicies: []policy.Policy{
+				genericarmclient.NewUserAgentPolicy(userAgent),
+			},
+		},
+	}
+
+	pipeline, err := armruntime.NewPipeline("generic", version.BuildVersion, creds, azruntime.PipelineOptions{}, opts)
+	return &pipeline, err
 }
