--- v2/tools/asoctl/internal/importing/importer_test.go
@@ -1,6 +1,0 @@
-/*
- * Copyright (c) Microsoft Corporation.
- * Licensed under the MIT license.
- */
-
-package importing
