+++ v2/tools/asoctl/internal/importing/resource_import_result.go
@@ -1,0 +1,68 @@
+/*
+ * Copyright (c) Microsoft Corporation.
+ * Licensed under the MIT license.
+ */
+
+package importing
+
+import (
+	"bufio"
+	"io"
+	"os"
+
+	"github.com/Azure/azure-service-operator/v2/pkg/genruntime"
+	"github.com/pkg/errors"
+	"gopkg.in/yaml.v3"
+)
+
+// ResourceImportResult represents the result of an import operation
+type ResourceImportResult struct {
+	resources []genruntime.MetaObject
+}
+
+func (r *ResourceImportResult) SaveToWriter(destination io.Writer) error {
+	buf := bufio.NewWriter(destination)
+	defer func(buf *bufio.Writer) {
+		_ = buf.Flush()
+	}(buf)
+
+	buf.WriteString("---\n")
+	for _, resource := range r.resources {
+		data, err := yaml.Marshal(resource)
+		if err != nil {
+			return errors.Wrap(err, "unable to save to writer")
+		}
+
+		buf.Write(data)
+		buf.WriteString("---\n")
+	}
+
+	return nil
+}
+
+func (r *ResourceImportResult) SaveToFile(filepath string) error {
+	file, err := os.Create(filepath)
+	if err != nil {
+		return errors.Wrapf(err, "unable to create file %s", filepath)
+	}
+
+	defer func() {
+		file.Close()
+
+		// if we are panicking, the file will be in a broken
+		// state, so remove it
+		if r := recover(); r != nil {
+			os.Remove(filepath)
+			panic(r)
+		}
+	}()
+
+	err = r.SaveToWriter(file)
+	if err != nil {
+		// cleanup in case of errors
+		file.Close()
+		os.Remove(filepath)
+	}
+
+	return errors.Wrapf(err, "unable to save to file %s", filepath)
+}
